splitting things out. removing lib-wrapper.
